Refactored file names, added enums, fixed indentations
