+++ python/torch_mlir_e2e_test/test_suite/__init__.py
@@ -41,9 +41,8 @@ def register_all_tests():
     from . import histogram_binning_calibration
     from . import rng
     from . import cast
-    from . import index_put
+    from . import scatter
     from . import pooling
     from . import return_types
     from . import control_flow
     from . import stats
-    from . import indirect_memory
